# [improve][broker] Added isActive in ManagedCursorImpl