@mattisonchao 2023-01-28
I am still wondering why we need `synchronized (activeCursors)`. It looks like `ManagedCursorContainer` is thread-safe.
@lhotari  Could you help to explain more about it? because I found you added it in https://github.com/apache/pulsar/pull/17273.