@lhotari 2023-01-30
> I am still wondering why we need `synchronized (activeCursors)`. It looks like `ManagedCursorContainer` is thread-safe. @lhotari Could you help to explain more about it? because I found you added it in #17273.

I guess I had added it to be consistent with the other method `deactivateCursor`, https://github.com/apache/pulsar/pull/17273/files#diff-f6a849bd8fdb782ef6c17a2e07a2c54c3dc7d1655c00ec3546d5f3b3fc61e970L3527  . That method contained `synchronized(activeCursors)`